stack/apikey-creator/pyproject.toml
[tool.poetry]
name = "apikey-creator"
version = "0.1.0"
description = ""
authors = ["Ben <ben@ahnlab.io>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
cryptography = "^43.0.1"
requests = "^2.32.3"
cfnresponse = "^1.1.5"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
boto3 = "^1.35.35"
cfnresponse = "^1.1.5"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
